+++ fastly/block_fastly_service_backend.go
@@ -6,6 +6,7 @@ import (
 	"log"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 
 	gofastly "github.com/fastly/go-fastly/v15/fastly"
 )
@@ -80,11 +81,25 @@ func (h *BackendServiceAttributeHandler) GetSchema() *schema.Schema {
 			Default:     200,
 			Description: "Maximum number of connections for this Backend. Default `200`",
 		},
+		"max_lifetime": {
+			Type:             schema.TypeInt,
+			Optional:         true,
+			Default:          0,
+			Description:      "Maximum time from creation (in milliseconds) that a pooled HTTP keepalive connection will be eligible for reuse; 0 is treated as unlimited - which is the default behavior.",
+			ValidateDiagFunc: validation.ToDiagFunc(validation.IntAtLeast(0)),
+		},
 		"max_tls_version": {
 			Type:        schema.TypeString,
 			Optional:    true,
 			Default:     "",
 			Description: "Maximum allowed TLS version on SSL connections to this backend.",
+		},
+		"max_use": {
+			Type:             schema.TypeInt,
+			Optional:         true,
+			Default:          0,
+			Description:      "Maximum number of requests allowed over a single, pooled HTTP keepalive connection to this backend; 0 is treated as unlimited - which is the default behavior.",
+			ValidateDiagFunc: validation.ToDiagFunc(validation.IntAtLeast(0)),
 		},
 		"min_tls_version": {
 			Type:        schema.TypeString,
@@ -317,6 +332,13 @@ func (h *BackendServiceAttributeHandler) buildCreateBackendInput(service string,
 		opts.KeepAliveTime = gofastly.ToPointer(resource["keepalive_time"].(int))
 	}
 
+	if resource["max_lifetime"].(int) > 0 {
+		opts.MaxLifetime = gofastly.ToPointer(resource["max_lifetime"].(int))
+	}
+	if resource["max_use"].(int) > 0 {
+		opts.MaxUse = gofastly.ToPointer(resource["max_use"].(int))
+	}
+
 	// WARNING: The following fields shouldn't have an empty string passed.
 	// As it will cause the Fastly API to return an error.
 	// This is because go-fastly v7+ will not 'omitempty' due to pointer type.
@@ -381,6 +403,12 @@ func (h *BackendServiceAttributeHandler) buildUpdateBackendInput(serviceID strin
 	}
 	if v, ok := modified["keepalive_time"]; ok {
 		opts.KeepAliveTime = gofastly.ToPointer(v.(int))
+	}
+	if v, ok := modified["max_lifetime"]; ok {
+		opts.MaxLifetime = gofastly.ToPointer(v.(int))
+	}
+	if v, ok := modified["max_use"]; ok {
+		opts.MaxUse = gofastly.ToPointer(v.(int))
 	}
 	if v, ok := modified["max_conn"]; ok {
 		opts.MaxConn = gofastly.ToPointer(v.(int))
@@ -478,6 +506,12 @@ func flattenBackend(remoteState []*gofastly.Backend, sa ServiceMetadata) []map[s
 		}
 		if resource.MaxConn != nil {
 			data["max_conn"] = *resource.MaxConn
+		}
+		if resource.MaxLifetime != nil {
+			data["max_lifetime"] = *resource.MaxLifetime
+		}
+		if resource.MaxUse != nil {
+			data["max_use"] = *resource.MaxUse
 		}
 		if resource.MaxTLSVersion != nil {
 			data["max_tls_version"] = *resource.MaxTLSVersion